+++ src/main/java/cn/xidian/meetingroom/model/Notification.java
@@ -5,8 +5,8 @@ import java.util.Date;
 
 @Data
 public class Notification {
-    private int notificationId;
-    private int userId;
+    private Long notificationId;
+    private Long userId;
     private String type;
     private String content;
     private boolean isRead;
